+++ src/datepicker/Month.svelte
@@ -79,8 +79,8 @@
     {#each [0] as item (legacy ? item : year)}
       <div
         class="grid-cell"
-        in:fly={{ x: direction * 50, duration: 200, delay: 80 }}
-        out:fade={{ duration: direction === 0 ? 0 : 160 }}
+        in:fly|local={{ x: direction * 50, duration: 200, delay: 80 }}
+        out:fade|local={{ duration: direction === 0 ? 0 : 160 }}
       >
         <div class="title" tabindex="0" on:keydown={onKeydown} on:click={onYear}>
           {("000" + year).slice(-4)}