+++ src/index.js
@@ -8,6 +8,8 @@ export {default as ExpansionPanel} from './ExpansionPanel.svelte';
 export {default as Icon} from './Icon.svelte';
 export {default as Menu} from './Menu.svelte';
 export {default as Menuitem} from './Menuitem.svelte';
+export {default as Modal} from './Modal.svelte';
+export {default as Popover} from './Popover.svelte';
 export {default as Radio} from './Radio.svelte';
 export {default as Ripple} from './Ripple.svelte';
 export {default as Sidepanel} from './Sidepanel.svelte';